orchestration/requirements.txt
apache-airflow-providers-apache-hdfs==4.11.1
apache-airflow-providers-apache-spark==5.4.1
apache-airflow-providers-common-compat==1.10.1
confluent-kafka==2.3.0